Use correct datatype for blobId, long (int64)

+++ CogniteSdk.Types/3D/3DListOutputsQuery.cs
@@ -52,6 +52,6 @@ namespace CogniteSdk
         /// Reference to 3D file containing output.
         /// 3D file can either be a single file or folder.
         /// </summary>
-        public int BlobId { get; set; }
+        public long BlobId { get; set; }
     }
 }
